# Use bits.OnesCount64 when available