--- a/AUTHORS.md
+++ b/AUTHORS.md
@@ -1,3 +1,4 @@
 ## AUTHORS
+- [Dominik Honnef](https://github.com/dominikh) dominik@honnef.co
 - [Dong-hee Na](https://github.com/corona10/) donghee.na92@gmail.com
 - [Gustavo Brunoro](https://github.com/brunoro/) git@hitnail.net