@dominikh 2018-05-02
@corona10 Is the file sorted in alphabetical order or in order of first contribution?